+++ packages/react-ui-kit/src/Misc/index.ts
@@ -26,3 +26,5 @@ export * from './Pagination';
 export * from './Pill';
 export * from './useLongTouch';
 export * from './useTimeout';
+export * from './Breadcrumbs';
+export * from './Badge';